requirements.txt
pytorch>=2.5.1+cu118
torchvision>=0.20.1+cu118
pillow>=11.0.0
selenium
matplotlib==3.3.4
tqdm
torchinfo